+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## mission1 - Controller created 21/10/2022
+
+- The controllers fibonacci is created
+  - Added the route api/v1/post/fibonacci responding with POST
+  - Added the base and fibonacci controller
+  - Added request and controller specs
+
 ## mission1 - Initial configuration 21/10/2022
 
 - Bundle and first commit